examples: move lib capture example to libcapture
To keep the simple example simple, move the lib based capture method
example to its own example.
