+++ examples/lib/libcapture/.gitignore
@@ -1,0 +1,2 @@
+!/Makefile.example.in
+Makefile.example
