+++ examples/lib/simple/main.c
@@ -16,62 +16,14 @@
  */
 
 #include "suricata.h"
-#include "conf.h"
-#include "pcap.h"
-#include "runmode-lib.h"
-#include "source-lib.h"
-#include "threadvars.h"
-
-/* Suricata worker thread in library mode.
-   The functions should be wrapped in an API layer. */
-void *suricataSimpleWorker(void *arg)
-{
-    char *pcap_file = (char *)arg;
-
-    /* Create worker. */
-    ThreadVars *tv = RunModeCreateWorker();
-    if (!tv) {
-        pthread_exit(NULL);
-    }
-
-    /* Start worker. */
-    if (RunModeSpawnWorker(tv) != 0) {
-        pthread_exit(NULL);
-    }
-
-    /* Replay pcap. */
-    pcap_t *fp = pcap_open_offline(pcap_file, NULL);
-    if (fp == NULL) {
-        pthread_exit(NULL);
-    }
-
-    int datalink = pcap_datalink(fp);
-    struct pcap_pkthdr pkthdr;
-    const u_char *packet;
-    while ((packet = pcap_next(fp, &pkthdr)) != NULL) {
-        if (TmModuleLibHandlePacket(tv, packet, datalink, pkthdr.ts, pkthdr.len, 0, 0, NULL) != 0) {
-            pthread_exit(NULL);
-        }
-    }
-    pcap_close(fp);
-
-    /* Cleanup. */
-    RunModeDestroyWorker(tv);
-    pthread_exit(NULL);
-}
 
 int main(int argc, char **argv)
 {
     SuricataPreInit(argv[0]);
 
     /* Parse command line options. This is optional, you could
-     * directly configure Suricata through the Conf API.
-       The last argument is the PCAP file to replay. */
-    SCParseCommandLine(argc - 1, argv);
-
-    /* Set lib runmode. There is currently no way to set it via
-       the Conf API. */
-    SuricataSetLibRunmode();
+     * directly configure Suricata through the Conf API. */
+    SCParseCommandLine(argc, argv);
 
     /* Validate/finalize the runmode. */
     if (SCFinalizeRunMode() != TM_ECODE_OK) {
@@ -95,27 +47,12 @@ int main(int argc, char **argv)
         exit(EXIT_FAILURE);
     }
 
-    /* Set "offline" runmode to replay a pcap in library mode. */
-    if (!ConfSetFromString("runmode=offline", 1)) {
-        exit(EXIT_FAILURE);
-    }
-
     SuricataInit();
-
-    /* Create and start worker on its own thread, passing the PCAP file
-       as argument. This needs to be done in between SuricataInit and
-       SuricataPostInit. */
-    pthread_t worker;
-    if (pthread_create(&worker, NULL, suricataSimpleWorker, argv[argc - 1]) != 0) {
-        exit(EXIT_FAILURE);
-    }
-
-    /* Need to introduce a little sleep to allow the worker thread to
-       initialize before SuricataPostInit invokes TmThreadContinueThreads().
-       This should be handle at the API level. */
-    usleep(100);
-
     SuricataPostInit();
+
+    /* Suricata is now running, but we enter a loop to keep it running
+     * until it shouldn't be running anymore. */
+    SuricataMainLoop();
 
     /* Shutdown engine. */
     SuricataShutdown();
